Added ability to browse through instagram posts
I used the tutorial here: https://medium.com/swlh/tutorial-web-scraping-instagrams-most-precious-resource-corgis-235bf0389b0c as a source of inspiration. The script can work very, very quickly to scrape url links of posts(instagram). Each iteration of the while loop corresponds to 30 posts, I plan to add a self.depth functionality to ask the user what the depth of the search will be.

The next step is to implement a set of keywords and then make the script search through posts and then only write the posts that contain predetermined keywords. I plan to implement a new loop for the urls gathered. I also need to study inner text functionalities of Selenium for keyword search. I might just write a Javascript script for this reason though...